refactor: (gallery) move SamplePages to Samples namespace and folder
