+++ source/iNKORE.UI.WPF.Modern.Gallery/Pages/Controls/Windows/SystemBackdropsPage.xaml.cs
@@ -2,7 +2,7 @@ using iNKORE.UI.WPF.Modern.Controls;
 using iNKORE.UI.WPF.Modern.Controls.Helpers;
 using iNKORE.UI.WPF.Modern.Controls.Primitives;
 using iNKORE.UI.WPF.Modern.Helpers.Styles;
-using iNKORE.UI.WPF.Modern.Gallery.SamplePages;
+using iNKORE.UI.WPF.Modern.Gallery.Samples;
 using System;
 using System.Collections.Generic;
 using System.Linq;
