+++ source/iNKORE.UI.WPF.Modern.Gallery/Samples/SampleContentDialogWindow.xaml
@@ -1,5 +1,5 @@
 <Window
-    x:Class="iNKORE.UI.WPF.Modern.Gallery.SamplePages.SampleContentDialogWindow"
+    x:Class="iNKORE.UI.WPF.Modern.Gallery.Samples.SampleContentDialogWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
