+++ source/iNKORE.UI.WPF.Modern.Gallery/Samples/SampleSystemBackdropsWindow.xaml
@@ -1,5 +1,5 @@
 <Window
-    x:Class="iNKORE.UI.WPF.Modern.Gallery.SamplePages.SampleSystemBackdropsWindow"
+    x:Class="iNKORE.UI.WPF.Modern.Gallery.Samples.SampleSystemBackdropsWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
